requirements.txt
matplotlib==3.3.4
pillow==9.1.0
tensorflow==2.6.2
opencv-python==4.2.0.34
numpy==1.19.5
pandas==1.1.5
keras==2.6.0
